+++ kotlin/src/main/kotlin/org/jdbi/v3/core/kotlin/KotlinMapper.kt
@@ -14,6 +14,7 @@
 package org.jdbi.v3.core.kotlin
 
 import org.jdbi.v3.core.annotation.internal.JdbiAnnotations
+import org.jdbi.v3.core.kotlin.internal.toJavaType
 import org.jdbi.v3.core.mapper.Nested
 import org.jdbi.v3.core.mapper.PropagateNull
 import org.jdbi.v3.core.mapper.RowMapper
@@ -192,16 +193,7 @@ class KotlinMapper(val kClass: KClass<*>, private val prefix: String = "") : Row
             val parameterName = addPropertyNamePrefix(prefix, parameter.paramName())
             val columnIndex = findColumnIndex(parameterName, columnNames, columnNameMatchers) { parameter.name }
             if (columnIndex.isPresent) {
-                val parameterType = parameter.type
-                // value types have a classifier that is different from the javaType. See a classifier is
-                // present and if yes, use that java type.
-                val javaType = if (parameterType.classifier != null && parameterType.classifier is KClass<*>) {
-                    (parameterType.classifier!! as KClass<*>).java
-                } else {
-                    parameterType.javaType
-                }
-
-                val type = QualifiedType.of(javaType)
+                val type = QualifiedType.of(toJavaType(parameter.type))
                     .withAnnotations(getQualifiers(parameter))
 
                 return ctx.findColumnMapperFor(type)
@@ -319,7 +311,7 @@ class KotlinMapper(val kClass: KClass<*>, private val prefix: String = "") : Row
 
     private data class ParamData(val type: ParamResolution, val mapper: RowMapper<*>?, val propagateNull: Boolean)
 
-    override fun toString() = "KotlinMapper(kClass=$kClass, prefix='$prefix')"
+    override fun toString() = "KotlinMapper(kClass=${kClass.qualifiedName}, prefix=$prefix)"
 
     private inner class BoundKotlinMapper(
         private val resolvedConstructorParameters: Map<KParameter, ParamData>,
@@ -363,6 +355,8 @@ class KotlinMapper(val kClass: KClass<*>, private val prefix: String = "") : Row
                 }
             }
         }
+
+        override fun toString() = "BoundKotlinMapper(kClass=${kClass.qualifiedName}, prefix=$prefix)"
     }
 }