+++ kotlin/src/main/kotlin/org/jdbi/v3/core/kotlin/KotlinPlugin.kt
@@ -14,20 +14,22 @@
 package org.jdbi.v3.core.kotlin
 
 import org.jdbi.v3.core.Jdbi
+import org.jdbi.v3.core.kotlin.internal.KotlinValueClassArgumentFactory
+import org.jdbi.v3.core.kotlin.internal.KotlinValueClassColumnMapperFactory
 import org.jdbi.v3.core.mapper.RowMappers
 import org.jdbi.v3.core.spi.JdbiPlugin
 
 /**
- * Installs Kotlin specific functionality.
+ * Installs Kotlin-specific functionality.
  *
- * <ul>
- *     <li>Kotlin inference interceptor that handles the {@link KotlinMapper} for registered types</li>
- *     <li>Kotlin mapper factory that creates implicit mappers for any data class</li>
- *     <li>Support for using Handles in Coroutines</li>
- * <ul>
+ *  * Kotlin inference interceptor that handles the [KotlinMapper] for registered types.
+ *  * Kotlin mapper factory that creates implicit mappers for any data class.
+ *  * Kotlin argument factory that binds Kotlin value classes as arguments.
+ *  * Kotlin column mapper factory that maps Kotlin value classes from result sets.
+ *  * Support for using Handles in Coroutines.
  *
- *     @property installKotlinMapperFactory If true, install the {@link KotlinMapperFactory}.
- *     @property enableCoroutineSupport If true, enable support for Kotlin Coroutines.
+ *   @property installKotlinMapperFactory If true, install support for data and value classes. The default is `true`.
+ *   @property enableCoroutineSupport If true, enable support for Kotlin Coroutines. The default is `false`.
  */
 class KotlinPlugin(private val installKotlinMapperFactory: Boolean = true, private val enableCoroutineSupport: Boolean = false) : JdbiPlugin.Singleton() {
 
@@ -38,12 +40,14 @@ class KotlinPlugin(private val installKotlinMapperFactory: Boolean = true, priva
 
         if (installKotlinMapperFactory) {
             jdbi.registerRowMapper(KotlinMapperFactory())
+            jdbi.registerColumnMapper(KotlinValueClassColumnMapperFactory())
+            jdbi.registerArgument(KotlinValueClassArgumentFactory())
         }
 
         // install a special handle scope that can deal with coroutines.
         //
         // Note that this needs to be revisited in the future if we move from the ThreadLocal in Jdbi
-        // to e.g. structured concurrency. Right now, this delegates to a ThreadLocalHandleScope.
+        // to e.g., structured concurrency. Right now, this delegates to a ThreadLocalHandleScope.
         if (enableCoroutineSupport) {
             jdbi.handleScope = CoroutineHandleScope()
         }