+++ kotlin/src/main/kotlin/org/jdbi/v3/core/kotlin/internal/KotlinPropertyArguments.kt
@@ -26,7 +26,6 @@ import kotlin.reflect.KMutableProperty1
 import kotlin.reflect.KProperty1
 import kotlin.reflect.full.memberProperties
 import kotlin.reflect.full.primaryConstructor
-import kotlin.reflect.jvm.javaType
 
 class KotlinPropertyArguments(obj: Any, prefix: String = "") : ObjectPropertyNamedArgumentFinder(prefix, obj) {
 
@@ -39,7 +38,7 @@ class KotlinPropertyArguments(obj: Any, prefix: String = "") : ObjectPropertyNam
     override fun getValue(name: String, ctx: StatementContext): Optional<TypedValue> {
         val property: KProperty1<*, *> = properties[name] ?: return Optional.empty()
         val mutableProperty = property as? KMutableProperty1
-        val type = QualifiedType.of(property.returnType.javaType)
+        val type = QualifiedType.of(toJavaType(property.returnType))
             .withAnnotations(
                 getQualifiers(
                     kClass.primaryConstructor?.run { parameters.find { it.name == name } },
@@ -55,5 +54,5 @@ class KotlinPropertyArguments(obj: Any, prefix: String = "") : ObjectPropertyNam
 
     override fun getNestedArgumentFinder(obj: TypedValue): NamedArgumentFinder = KotlinPropertyArguments(obj.value)
 
-    override fun toString() = obj.toString()
+    override fun toString() = "KotlinPropertyArguments(obj=${kClass.qualifiedName}, prefix=$prefix)"
 }