+++ kotlin/src/test/kotlin/org/jdbi/v3/core/kotlin/KotlinValueTypeTest.kt
@@ -17,23 +17,33 @@ import org.assertj.core.api.Assertions.assertThat
 import org.jdbi.v3.core.Handle
 import org.jdbi.v3.core.argument.AbstractArgumentFactory
 import org.jdbi.v3.core.argument.Argument
+import org.jdbi.v3.core.argument.ArgumentFactory
 import org.jdbi.v3.core.config.ConfigRegistry
+import org.jdbi.v3.core.kotlin.internal.KotlinValueClassArgumentFactory
+import org.jdbi.v3.core.kotlin.internal.KotlinValueClassColumnMapperFactory
 import org.jdbi.v3.core.mapper.ColumnMapper
+import org.jdbi.v3.core.mapper.ColumnMapperFactory
 import org.jdbi.v3.core.statement.StatementContext
 import org.jdbi.v3.testing.junit5.JdbiExtension
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.RegisterExtension
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
+import org.junit.jupiter.params.provider.MethodSource
+import java.lang.reflect.Type
 import java.sql.PreparedStatement
 import java.sql.ResultSet
 import java.sql.Types
+import java.util.Optional
+import java.util.stream.Stream
 
 class KotlinValueTypeTest {
 
     @RegisterExtension
     @JvmField
-    val h2Extension: JdbiExtension = JdbiExtension.h2().withPlugin(KotlinPlugin())
-        .withInitializer { ds, handle -> handle.execute("CREATE TABLE something (id integer, first text)") }
+    val h2Extension: JdbiExtension = JdbiExtension.h2().withPlugin(KotlinPlugin(installKotlinMapperFactory = false))
+        .withInitializer { ds, handle -> handle.execute("CREATE TABLE something (id integer, first text, other text)") }
 
     private lateinit var handle: Handle
 
@@ -46,7 +56,7 @@ class KotlinValueTypeTest {
     value class MagicType(val first: String)
 
     class MagicTypeArgumentFactory : AbstractArgumentFactory<MagicType>(Types.VARCHAR) {
-        override fun build(value: MagicType?, config: ConfigRegistry): Argument? =
+        override fun build(value: MagicType?, config: ConfigRegistry): Argument =
             Argument { position: Int, statement: PreparedStatement, ctx: StatementContext ->
                 if (value != null) {
                     statement.setString(position, value.first)
@@ -63,10 +73,44 @@ class KotlinValueTypeTest {
         }
     }
 
-    @Test
-    fun testValueClass() {
-        handle.registerArgument(MagicTypeArgumentFactory())
-        handle.registerColumnMapper(MagicTypeColumnMapper())
+    class MagicTypeColumnMapperFactory : ColumnMapperFactory {
+        override fun build(type: Type?, config: ConfigRegistry?): Optional<ColumnMapper<*>> {
+            if (type == MagicType::class.java) {
+                return Optional.of(MagicTypeColumnMapper())
+            }
+            return Optional.empty()
+        }
+    }
+
+    companion object {
+        @JvmStatic
+        fun arguments(): Stream<Arguments> {
+            val mappers = listOf(
+                KotlinValueClassColumnMapperFactory(), // Kotlin Value class support
+                MagicTypeColumnMapperFactory()
+            )
+            val argumentFactories = listOf(
+                KotlinValueClassArgumentFactory(), // Kotlin Value class support
+                MagicTypeArgumentFactory()
+            )
+
+            val arguments = ArrayList<Arguments>()
+
+            for (mapper in mappers) {
+                for (argumentFactory in argumentFactories) {
+                    arguments.add(Arguments.of(mapper, argumentFactory))
+                }
+            }
+
+            return arguments.stream()
+        }
+    }
+
+    @ParameterizedTest
+    @MethodSource("arguments")
+    fun testValueClass(columnMapperFactory: ColumnMapperFactory, argumentFactory: ArgumentFactory) {
+        handle.registerArgument(argumentFactory)
+        handle.registerColumnMapper(columnMapperFactory)
         val expected = MagicType("does this work?")
 
         handle.createUpdate("INSERT INTO something(id, first) VALUES(:id, :first)")
@@ -81,12 +125,59 @@ class KotlinValueTypeTest {
         assertThat(result).isEqualTo(expected)
     }
 
+    @ParameterizedTest
+    @MethodSource("arguments")
+    fun testValueClassList(columnMapperFactory: ColumnMapperFactory, argumentFactory: ArgumentFactory) {
+        handle.registerArgument(argumentFactory)
+        handle.registerColumnMapper(columnMapperFactory)
+        val expected = MagicType("does this work?")
+
+        for (i in 1..10) {
+            handle.createUpdate("INSERT INTO something(id, first) VALUES(:id, :first)")
+                .bind("id", i)
+                .bind("first", expected)
+                .execute()
+        }
+
+        val result = handle.createQuery("SELECT first FROM something ORDER BY id")
+            .mapTo<MagicType>()
+            .list()
+
+        assertThat(result).hasSize(10)
+        for (i in 0..9) {
+            assertThat(result.get(i)).isEqualTo(expected)
+        }
+    }
+
+    @ParameterizedTest
+    @MethodSource("arguments")
+    fun testValueClassNull(columnMapperFactory: ColumnMapperFactory, argumentFactory: ArgumentFactory) {
+        handle.registerArgument(argumentFactory)
+        handle.registerColumnMapper(columnMapperFactory)
+
+        handle.createUpdate("INSERT INTO something(id, first) VALUES(:id, :first)")
+            .bind("id", 1)
+            .bindNull("first", Types.VARCHAR)
+            .execute()
+
+        val result = handle.createQuery("SELECT first FROM something")
+            .mapTo<MagicType>()
+            .single()
+
+        assertThat(result).isNull()
+    }
+
     data class TheThings(val id: Int, val first: MagicType)
 
-    @Test
-    fun testValueBean() {
-        handle.registerArgument(MagicTypeArgumentFactory())
-        handle.registerColumnMapper(MagicTypeColumnMapper())
+    @ParameterizedTest
+    @MethodSource("arguments")
+    fun testValueBean(columnMapperFactory: ColumnMapperFactory, argumentFactory: ArgumentFactory) {
+        handle.registerArgument(argumentFactory)
+        handle.registerColumnMapper(columnMapperFactory)
+
+        // Not automatically registered b/c using installKotlinMapperFactory = false
+        handle.registerRowMapper(KotlinMapperFactory())
+
         val expected = MagicType("does this work?")
 
         handle.createUpdate("INSERT INTO something(id, first) VALUES(:id, :first)")
@@ -99,5 +190,61 @@ class KotlinValueTypeTest {
             .single()
 
         assertThat(result).isEqualTo(TheThings(1, expected))
+    }
+
+    @ParameterizedTest
+    @MethodSource("arguments")
+    fun testValueBeanList(columnMapperFactory: ColumnMapperFactory, argumentFactory: ArgumentFactory) {
+        handle.registerArgument(argumentFactory)
+        handle.registerColumnMapper(columnMapperFactory)
+
+        // Not automatically registered b/c using installKotlinMapperFactory = false
+        handle.registerRowMapper(KotlinMapperFactory())
+
+        val expected = MagicType("does this work?")
+
+        for (i in 1..10) {
+            handle.createUpdate("INSERT INTO something(id, first) VALUES(:id, :first)")
+                .bind("id", i)
+                .bind("first", expected)
+                .execute()
+        }
+
+        val result = handle.createQuery("SELECT id, first from something")
+            .mapTo<TheThings>()
+            .list()
+
+        assertThat(result).hasSize(10)
+        for (i in 0..9) {
+            assertThat(result.get(i))
+                .hasFieldOrPropertyWithValue("id", i + 1)
+                .hasFieldOrPropertyWithValue("first", expected.first) // someone brought a java library to a kotlin test...
+        }
+    }
+
+    @JvmInline
+    value class SpaceType(val value: String)
+
+    data class MoreThings(val id: Int, val first: MagicType, val other: SpaceType)
+
+    @Test
+    fun testMultiValueClasses() {
+        handle.registerArgument(KotlinValueClassArgumentFactory())
+        handle.registerColumnMapper(KotlinValueClassColumnMapperFactory())
+        handle.registerRowMapper(KotlinMapperFactory())
+
+        val expected = MagicType("does this work?")
+        val other = SpaceType("no it does not.")
+        val ktBean = MoreThings(1, expected, other)
+
+        handle.createUpdate("INSERT INTO something(id, first, other) VALUES(:id, :first, :other)")
+            .bindKotlin(ktBean)
+            .execute()
+
+        val result = handle.createQuery("SELECT id, first, other from something")
+            .mapTo<MoreThings>()
+            .single()
+
+        assertThat(result).isEqualTo(MoreThings(1, expected, other))
     }
 }